# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API",
        "description": "API UKM Politeknik Negeri Lampung",
        "contact": {
            "email": "leopradana013gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/newshit/{id}": {
            "get": {
                "tags": [
                    "News Hit"
                ],
                "summary": "Get Detail News Hit",
                "operationId": "ab49b961d49cc25b8119e0e4a83a0d7d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "_id": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "nama": {
                                            "type": "string",
                                            "example": "example nama"
                                        },
                                        "jenis": {
                                            "type": "enum",
                                            "example": "example jenis"
                                        },
                                        "singkatan_ukm": {
                                            "type": "string",
                                            "example": "example singkatan_ukm"
                                        },
                                        "foto_ukm": {
                                            "type": "text",
                                            "example": "example foto_ukm"
                                        },
                                        "keterangan": {
                                            "type": "string",
                                            "example": "example keterangan"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}